Update WGenerator.java
I really just shitted all over this file but in the end I think it turned out for the best. I added the functionality needed to load a heightmap from file as well as edited a bunch of windows parameters and layouts to get my buttons to fit. The sizes of the various windows elements are not guaranteed to have any logical meaning behind them.

Also anytime an action completes it calls a system beep because the amount of times I hit drop dirt and then got distracted on the internet and forgot about it in the background is too damn high.